Move pytree-specific visitors to appropriate subdirectory
