+++ yapf/yapflib/format_decision_state.py
@@ -26,9 +26,9 @@ through the code to commit the whitespace formatting.
   FormatDecisionState: main class exported by this module.
 """
 
+from yapf.pytree import split_penalty
 from yapf.yapflib import logical_line
 from yapf.yapflib import object_state
-from yapf.yapflib import split_penalty
 from yapf.yapflib import style
 from yapf.yapflib import subtypes
