+++ yapf/yapflib/logical_line.py
@@ -20,9 +20,9 @@ perform the wrapping required to comply with the style guide.
 """
 
 from yapf.pytree import pytree_utils
+from yapf.pytree import split_penalty
 from yapf.yapflib import format_token
 from yapf.yapflib import py3compat
-from yapf.yapflib import split_penalty
 from yapf.yapflib import style
 from yapf.yapflib import subtypes
