+++ yapftests/format_decision_state_test.py
@@ -17,6 +17,7 @@ import textwrap
 import unittest
 
 from yapf.pytree import pytree_utils
+
 from yapf.yapflib import format_decision_state
 from yapf.yapflib import logical_line
 from yapf.yapflib import style
