+++ yapftests/logical_line_test.py
@@ -19,9 +19,9 @@ import unittest
 from lib2to3 import pytree
 from lib2to3.pgen2 import token
 
+from yapf.pytree import split_penalty
 from yapf.yapflib import format_token
 from yapf.yapflib import logical_line
-from yapf.yapflib import split_penalty
 
 from yapftests import yapf_test_helper
