+++ yapftests/split_penalty_test.py
@@ -21,7 +21,7 @@ from lib2to3 import pytree
 
 from yapf.pytree import pytree_utils
 from yapf.pytree import pytree_visitor
-from yapf.yapflib import split_penalty
+from yapf.pytree import split_penalty
 from yapf.yapflib import style
 
 from yapftests import yapf_test_helper
